backend/requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
